fixing a bunch of bugs with google cloud, updates to readme, backend has locks to prevent spamming requests

See merge request purtilo/radiology!31

+++ platform/broker/src/orthanc_functions.py
@@ -37,7 +37,7 @@ def change_tags(series_UID):#make sure this is an RTstruct series
 
 
 def get_first_dicom_image_series_from_study(patient_id, study_UID, save_directory):
-    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc')
+    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc', timeout=60)
 
     patient_file_id = pyorthanc.find_patients(orthanc,query={'PatientID': patient_id})[0]#there should only be one
     patient = pyorthanc.Patient(patient_file_id.get_main_information()["ID"],orthanc)
@@ -70,11 +70,11 @@ def get_first_dicom_image_series_from_study(patient_id, study_UID, save_director
     image_folder = os.listdir(study_dir)[image_order_num]
     return first_folder, study_dir + "/"+image_folder
 
-def get_dicom_series_by_id(series_instance_uid, save_directory):
+def get_dicom_series_by_id(series_instance_uid, save_directory, series_obj_out=[], extract_zip=True):
     print('saving', save_directory)
     url = orthanc_url
 
-    orthanc = pyorthanc.Orthanc(url, username='orthanc', password='orthanc')
+    orthanc = pyorthanc.Orthanc(url, username='orthanc', password='orthanc', timeout=60)
     valid_series = pyorthanc.find_series(orthanc, query={'SeriesInstanceUID': series_instance_uid})
     if len(valid_series) == 0:
         raise Exception('No series found with UID:', series_instance_uid)
@@ -87,19 +87,31 @@ def get_dicom_series_by_id(series_instance_uid, save_directory):
     os.makedirs(save_directory, exist_ok=True)
     zip_path = os.path.join(save_directory, 'series.zip')
     a_series.download(zip_path, with_progres=False)
-    # NOTE probably just unzip it with command line
-    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
-        # zip_ref.extractall(zip_path, [r'PANCREAS_0005 PANCREAS_0005/'])
-        zip_ref.extractall(save_directory)
-        os.remove(zip_path)
+    if extract_zip:
+        with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+            # zip_ref.extractall(zip_path, [r'PANCREAS_0005 PANCREAS_0005/'])
+            zip_ref.extractall(save_directory)
+            os.remove(zip_path)
     
     patient = a_series.parent_patient 
+    if isinstance(series_obj_out, list):
+        series_obj_out.append(a_series)
+        # a_series.parent_study.uid
+        # a_series.description
     
-    return os.path.join(save_directory, f'{patient.patient_id} {patient.name}')
+    return os.path.join(save_directory, f'{patient.patient_id} {patient.name}') if extract_zip else zip_path
 
+def extract_dicom_series_zip(zip_path, save_directory, remove_original=False):
+    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+            zip_ref.extractall(save_directory)
+            if remove_original:
+                os.remove(zip_path)
+    
+    return save_directory
 
 def uploadSegFile(file_path, remove_original=False):
-    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc')
+    print('Uploading', file_path, 'to orthanc...')
+    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc', timeout=60)
     with open(file_path, 'rb') as file:
         orthanc.post_instances(file.read())
 
@@ -109,7 +121,7 @@ def uploadSegFile(file_path, remove_original=False):
             os.remove(file_path)
             
 def get_modality_of_series(series_UID):
-    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc')
+    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc', timeout=60)
     series = None
     try:
         series = pyorthanc.find_series(orthanc, query={"SeriesInstanceUID":series_UID})[0]#Should only be one
@@ -118,3 +130,39 @@ def get_modality_of_series(series_UID):
     
     series = series.get_main_information()
     return series["MainDicomTags"]["Modality"]
+
+def get_next_available_iterative_name_for_series(base_series_name, parent_study_uid, split_char='_', modality='SEG'):
+    orthanc = pyorthanc.Orthanc(orthanc_url, username='orthanc', password='orthanc', timeout=60)
+    valid_studies = pyorthanc.find_studies(orthanc, query={'StudyInstanceUID': parent_study_uid})
+    if len(valid_studies) == 0:
+        raise Exception('Could not find any studies with UID', parent_study_uid)
+    # print(valid_studies[0])
+    # print([s.modality for s in valid_studies[0].series])
+    # print([s.description for s in valid_studies[0].series])
+    # print([split_char.join(s.description.split(split_char)[:-1]) for s in valid_studies[0].series])
+    intersecting_series = []
+    for s in valid_studies[0].series:
+        try:
+            if s.modality == modality and \
+               s.description.count(split_char) >= 1 and \
+               split_char.join(s.description.split(split_char)[:-1]) == base_series_name:
+                   intersecting_series.append(s.description)
+        except Exception as e:
+            continue
+                
+    # intersecting_series = [s.description for s in valid_studies[0].series if (
+    #     s.modality==modality and 
+    #     s.description.count(split_char) >= 1 and
+    #     split_char.join(s.description.split(split_char)[:-1])==base_series_name
+    #     # s.description.startswith(base_series_name)
+    #     )]
+    # print(intersecting_series)
+    # intersecting_series.sort(key=lambda s: s.description)
+    next_valid_iteration = 1
+    for i in range(1, len(intersecting_series) + 2):
+        if f'{base_series_name}{split_char}{i}' not in intersecting_series:
+            next_valid_iteration = i
+            print('next valid iter:', next_valid_iteration)
+            break
+            
+    return f'{base_series_name}{split_char}{next_valid_iteration}'
